added junit tests

+++ src/main/java/com/techaccelarators/ifind/service/impl/BankServiceImpl.java
@@ -1,26 +1,28 @@
 package com.techaccelarators.ifind.service.impl;
 
 import com.techaccelarators.ifind.domain.Bank;
+import com.techaccelarators.ifind.domain.ServiceType;
 import com.techaccelarators.ifind.domain.enums.Status;
 import com.techaccelarators.ifind.dtos.bank.BankRequestDto;
-import com.techaccelarators.ifind.exception.InvalidRequestException;
-import com.techaccelarators.ifind.exception.RecordNotFoundException;
+import com.techaccelarators.ifind.exception.*;
 import com.techaccelarators.ifind.repository.BankRepository;
 import com.techaccelarators.ifind.service.BankService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 @Service
 @AllArgsConstructor
 @Slf4j
 public class BankServiceImpl implements BankService {
-
     private final BankRepository bankRepository;
     @Override
     public Bank createBank(BankRequestDto bankRequestDto) {
+        log.info("Creating a Bank with arguments: {}",bankRequestDto);
         checkUnique(bankRequestDto, null);
         Bank bank = Bank.builder()
                 .name(bankRequestDto.getName())
@@ -28,9 +30,9 @@ public class BankServiceImpl implements BankService {
         bank.setStatus(Status.ACTIVE);
         return bankRepository.save(bank);
     }
-
     @Override
     public Bank updateBank(Long id, BankRequestDto bankRequestDto) {
+        log.info("Updating a bank with an id: {}",id);
         Bank bank = getBankById(id);
         checkUnique(bankRequestDto, id);
         bank.setName(bankRequestDto.getName());
@@ -40,34 +42,36 @@ public class BankServiceImpl implements BankService {
 
         return bankRepository.save(bank);
     }
-
     @Override
     public Page<Bank> getAllBanks(Pageable pageable) {
+        log.info("Getting all available banks");
         return bankRepository.findAll(pageable);
     }
-
     @Override
     public Bank getBankById(Long id) {
+        log.info("Getting a bank with id: {}",id);
         return bankRepository.findById(id)
                 .orElseThrow(()-> new RecordNotFoundException("Bank Not Found"));
     }
-
     @Override
     public Bank getBankByName(String name) {
+        log.info("Getting a bank with name: {}",name);
         return bankRepository.findBankByName(name)
-                .orElseThrow(()-> new RecordNotFoundException("Bank Not Found"));
+                .orElseThrow(()->{
+                    log.info("Bank with name: {}  could not be found ",name);
+                    return new RecordNotFoundException("Bank Not Found");
+                });
     }
-
     @Override
     public Page<Bank> searchBank(String searchParam, Pageable pageable) {
         try {
             String searchWord = "%".concat(searchParam).concat("%");
             return bankRepository.findAllByNameLikeIgnoreCase(searchWord, pageable);
-        } catch (Exception ex) {
+        } catch (InvalidRequestException ex) {
+            log.info("Could not find the searched bank with name: {}",searchParam);
             throw new InvalidRequestException(ex.getMessage());
         }
     }
-
     @Override
     public Bank toggleBankStatus(Long id) {
         Bank bank = getBankById(id);
@@ -79,17 +83,28 @@ public class BankServiceImpl implements BankService {
         }
         return bankRepository.save(bank);
     }
-
     @Override
     public void deleteBank(Long id) {
-        Bank bank = getBankById(id);
-        bankRepository.delete(bank);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null || authentication.getPrincipal() == null) {
+            throw new UnauthorizedException("Unauthorized");
+        }
+        String role = authentication.getAuthorities().stream().findFirst().get().getAuthority();
+        if (!role.equals("ADMIN")) {
+            throw new ForbiddenException("Forbidden");
+        }
+        try {
+            Bank bank = getBankById(id);
+            bankRepository.delete(bank);
+        } catch (Exception e) {
+            throw new RuntimeException("Failed to delete the service type");
+        }
     }
     private void checkUnique(BankRequestDto request, Long id) {
         log.info("Checking bank uniqueness with name: {} and id: {}", request.getName(), id);
         bankRepository.findByNameIgnoreCase(request.getName())
                 .filter(bank -> !bank.getId().equals(id))
-                .ifPresent(customer -> {
+                .ifPresent(bank -> {
                     throw new InvalidRequestException("Bank with given name already exists");
                 });
     }
